spec update: Use u16 instead of u8
The funding_confirms_within_blocks should be a u8. This implies 0 <= funding_confirms_within_blocks <= 255.

The default-value used by the client in balance-of-satoshi is 288 blocks.
This means the client gives the server approximately 48 hours to ensure the funding the transaction is confirmed.

I've launched a proposal to update the spec but implement it here
anyway.

- See https://github.com/BitcoinAndLightningLayerSpecs/lsp/issues/112